Started creating a python equivalent of run_all.
[Details]
A new python-based equivalent of the run_all file is created in this
commit. The python-based run_all file should perform identically to the
bash script version of the run_all file. The differences are that the
python-based script can be run on any operating system (not dependant on
linux operating systems or systems with bash anymore), and that flitsr
can now be packaged as a purely python project.

[Note]
The run_all script in python is almost complete, with most components
working. The only components left are the summarization of the files
after flitsr has run.

[File's changed]
	new file:   run_all.py