Changed main methods for argparse handling.
[Details]
For all entry-points for flitsr, added an explicit main() method which
takes a mandatory argument with the options. The command-line entry
point now just calls this method. Also made sure all entry-points used
the arguments correctly, and that the command-line arguments in
sys.argv are correctly passed to argparse.

These changes are necessary for the new run_all script, which needs to
call a main method of an entry-point.

[File's changed]
	modified:   flitsr.py
	modified:   merge.py
	modified:   ordering_driver.py
	modified:   percent_at_n.py