Completed the python run_all script.
[Details]
Finished the implementation of the python-based run_all script. This
script should now do everything that the bash-based equivalent could. To
switch over, the run_all script is replaced with a call to the
python-based script, and the old bash-based script is stored in
old_run_all.

Also updated the README to reflect the change in command-line options.

[File's changed]
	modified:   README.md
	new file:   bin/old_run_all
	modified:   bin/run_all
	modified:   flitsr/run_all.py