Added input type enum & transform script.
[Details]
Added a new Enum called the InputType, which specifies the input
spectral format (TCM or GZoltar) used. This is used in the argument
parsing now instead of the "magic numbers" 0 and 1 for TCM and GZoltar
respectively.

Also added a new script called 'transform' which transforms a given
input file from one spectral format to another. This uses the
pre-existing print_tcm output function, and a new print_gzoltar output
function.

[File's changed]
	modified:   args.py
	modified:   flitsr.py
	new file:   input_type.py
	modified:   ordering_driver.py
	modified:   output.py
	modified:   spectrum.py
	new file:   transform.py