Working version of new advanced types.
[Details]
Almost finished the complete overhaul of the advanced types system. The
system now sucessfully detects advanced types in the flitsr.advanced
package, automatically compiles a list of their parameters, and
automatically links them in to be run.

[Note]
The implementation may still be broken when using the run_all evaluation
script.

[File's changed]
	modified:   advanced/__init__.py
	modified:   advanced/flitsr.py
	modified:   advanced/parallel.py
	modified:   args.py
	modified:   main.py