Removed dicts for advanced types, kept Enums.
[Details]
Removed the need to have both the dict for advanced types and their
classes AND the Enum of just the names of the advanced types. The dicts
are still created, but they are used locally to instantiate the Enums
with both the name of the advanced type, along with its class as the
Enum value. This makes the code dealing with advanced types in the rest
of the system much simpler.

[File's changed]
	modified:   __init__.py
	modified:   flitsr.py
	modified:   ../args.py
	modified:   ../main.py
	modified:   ../ordering_driver.py