Added advanced type custom output strings.
[Details]
Added the ability for an advanced type to have a custom output string.
By defining the _print_name class variable in an advanced type class,
the system will use this instead of the name of the class only when
generating the file name for printing out to a results file. This
option now allows the SBFL and FLITSR MULTI advanced types to behave
exactly the same as they did before the advanced types refactoring.

[File's changed]
	modified:   flitsr/advanced/__init__.py
	modified:   flitsr/advanced/flitsr.py
	modified:   flitsr/advanced/sbfl.py