Improved spectrum efficiency & bug fixes.
[Details]
Improved the effiency of spectrum operations in two ways:
1. Changed the spectrum components to use slotted attributes, which
   speeds up access and improves memory comsumption
2. Changed the SpectrumBuilder to only store execution information
   for elements actually executed in tests (and not non-executions).
   This was technically done in the input and tcm_input classes, but
   is now explicitly done by the SpectrumBuilder.

Also added the SpectrumUpdater class within the spectrumBuilder module,
which will be useful for updating an existing spectrum.

[Bugs]
Fixed a bug in the constructor of the spectrum from a SpectrumBuilder
which would increase counts for elements on tests that they weren't even
executed in.

[File's changed]
	modified:   flitsr/input.py
	modified:   flitsr/spectrum.py
	modified:   flitsr/spectrumBuilder.py
	modified:   flitsr/tcm_input.py