Added collect results option for run_all.
[Details]
Added a new option (-C) to the run_all script, which ONLY collects the
pre-generated flitsr outputs into results files, instead of running
flitsr and then collecting the outputs. This is useful for outputs that
are generated by flitsr outside of the run_all script, that are needing
to be collected for the merge script or other uses.

[File's changed]
	modified:   flitsr/run_all.py