Added advanced type decorators & fixed tiebrk.
[Details]
Added the 'attributes' module in the advanced package that defines a set
of decorators that advanced types can use to define certain attributes
for the type or its parameters. In particular the following was
changed/added:
1. Added the "@existing" decorator which can be used for the __init__
   function to specify a parameter that is not an advanced type option,
   but an existing flitsr option. No command line argument is added for
   this parameter, and instead, the existing flitsr option is given to
   the advanced type.
2. Changed the "_<parameter>_opts" field in an advanced type to the
   "@choices" decorator. This decorator can now be used for the
   __init__ function, to specify choices for a given parameter
3. Changed the "_print_name" field in an advanced type to the
   "@print_name" decorator. This decorator can now be used on the
   advanced type class to specify an alternate name to use when
   printing out results for this advanced type.
All of these new options are also handled in the flitsr "args" module as
well.

Also fixed up the use of tie-breaking options within FLITSR.
Tie-breaking options are now in the form of an Enum, which is used by
the Ranking class to rank. This option is available in the global
options and can be used (by the "@existing" decorator described above)
in any advanced type.

[File's changed]
	modified:   advanced/__init__.py
	modified:   advanced/artemis_wrapper.py
	new file:   advanced/attributes.py
	modified:   advanced/flitsr.py
	modified:   advanced/parallel.py
	modified:   advanced/sbfl.py
	modified:   advanced/sliced.py
	modified:   args.py
	modified:   ranking.py
	modified:   suspicious.py