Changes to the Sliced advanced type.
[Details]
Changed the Sliced advanced type to expect the base spectrum as the
usual FLITSR input, and the sliced spectrum as its input.

Also added the functionality for the sliced spectrum name to include
parts of the original input name and path. This makes it possible to run
the Sliced advanced type with the run_all script.

[Note]
In order to prevent arbitrary code from potentially being executed,
which is a security risk, the sliced spectrum name only has access to
the path, name and extension of the original input. This is somewhat
restrictive, however it does still enable the use of the run_all script
if file names are changed on the OS.

[File's changed]
	modified:   advanced/attributes.py
	modified:   advanced/sliced.py