Changed Sliced back to using other spectrum.
[Details]
Changed the Sliced advanced type back to using the sliced spectrum as
the FLITSR input, and the original spectrum as the parameter. The reason
for this is because this way around makes more sense as the user would
be running on the sliced spectrum.

I may change this back again :)

[File's changed]
	modified:   advanced/sliced.py