Modified mode printing order for merge script.
[Details]
Changed the order that modes are printed out in the results of the merge
script to now be independent of any common suffixes. This is useful for
comparison of two merged results files since modes contain the listed
names of the advanced types used, which may have common suffixes that
change the order of modes.

For example if two evaluations are run, one with the modes 'flitsr' and
'flitsr_multi', and another with the added advanced type 'sliced',
making the modes 'flitsr_sliced' and 'flitsr_multi_sliced'; in the
first one 'flitsr' would appear before 'flitsr_multi' in the results
file, but this would be swapped for the other evaluation. The reason
is because the 's' in 'flitsr_sliced' makes that mode "greater" than
'flitsr_multi_sliced'. By removing common suffixes this problem was
solved, so that the results files are predictably comparable.

[File's changed]
	modified:   merge.py