Fixed a bug in the SpectrumUpdater.
[Bugs]
The SpectrumUpdater class used groups and elements from multiple
spectra interchangably, however because of the way these are
implemented (using indexes as unique IDs), groups and elements from
one spectrum should NOT be used in another. To fix this, I refactored
the SpectrumUpdater and its methods to ensure that elements and groups
are only used for the spectrum they belong to. I also added a method to
check if two elements (from different spectra) are identical in the
Element class, which enables the SpectrumUpdater to work. Since
groups only make sense for their own spectrum, nothing was added to
them.

[File's changed]

	modified:   advanced/sliced.py
	modified:   spectrum.py
	modified:   spectrumBuilder.py