Finished implementing new input types in system.
[Details]
Finished adapting the rest of FLITSR to support the new input system.
This commit changes the ordering_driver and the transform scripts to use
the new input types system. In order to make the implementation more
modular and readable, the methods to print out the spectrum in each
input type (TCM and GZoltar) were moved to the corresponding new input
type classes, as the method 'write_spectrum'. This functionality is
optional for new input types (i.e. you do not need to implement
outputting a spectrum to an input type when adding a new input type),
but implementing this for new input types will allow spectral output
in that format.

The only places that use the new 'write_spectrum' methods are the
transform script and the parallel technique. The parallel technique only
prints out to TCM format, which is guaranteed to have the method
implemented, and the transform script is structured so that the user
should be aware why the program could not print to the given format if
the method is not implemented.

[File's changed]
	modified:   flitsr/advanced/parallel.py
	modified:   flitsr/input/__init__.py
	modified:   flitsr/input/gzoltar_input.py
	modified:   flitsr/input/input_reader.py
	modified:   flitsr/input/tcm_input.py
	deleted:    flitsr/input_type.py
	modified:   flitsr/ordering_driver.py
	modified:   flitsr/output.py
	modified:   flitsr/spectrum.py
	modified:   flitsr/transform.py