Enabled entry points for advanced types.
[Details]
Added the ability for FLITSR to discover advanced types implemented in
plugins by using entry points. Anyone creating a new advanced type for
FLITSR need only create a plugin that specifies the advanced type as an
entry point. Then to use it, the plugin must simply be installed in
the same environment as flitsr. This makes it easy to implement new
features for flitsr as plugins.

[Note]
Also changed FLITSR's argument handling to only include advanced type
groups (e.g. Refiner, Ranker, or Cluster) in FLITSR's arguments if
there are advanced types belonging to that group.

[File's changed]
	modified:   flitsr/advanced/__init__.py
	modified:   flitsr/args.py