Added "buckets" for spectral removal.
[Details]
Added the ability to store spectral tests and groups in buckets when
removing them from the spectrum. Before, all removed tests were stored
in a single cache, or "bucket", and removed groups were not stored. This
commit now gives the user of the spectrum the option to store
removed spectral tests and groups in custom buckets, which can allow for
multiple algorithms that modify the spectrum to remove things without
interfering with the other algorithms.

[Note]
The flitsr and parallel algorithms that already removed things from the
spectrum have been updated to respect the new functionality.

[File's changed]
	modified:   flitsr/advanced/flitsr.py
	modified:   flitsr/advanced/parallel.py
	modified:   flitsr/spectrum.py