Refactored Rankings to also take elements.
[Details]
Completely refactored all output functionality in FLITSR. The main
changes are that Rankings can now be made up of Groups OR Elements. In
addition, all aspects related to a spectrum are now removed from output
functionality.

The above two changes means that rankings do not have to refer to the
same Groups as any spectrum; Groups in a ranking can thus now be
modified, split, or even completely dismantled.

[File's changed]
	modified:   advanced/flitsr.py
	modified:   cutoff_points.py
	modified:   main.py
	modified:   ordering_driver.py
	modified:   output.py
	modified:   percent_at_n.py
	modified:   precision_recall.py
	modified:   ranking.py
	modified:   read_ranking.py
	modified:   spectrum.py
	modified:   spectrumBuilder.py
	modified:   split_faults.py
	modified:   tie.py