Various improvements to spectrum test removal.
[Details]
Added the ability to add back a single removed test from a spectrum.
Also made sure that removing a test is only done on tests that are
actually in the spectrum for the "remove_tests" method. The method
quietly returns if not (since the test is technically already
"removed" in that case).

[File's changed]
	modified:   flitsr/spectrum.py