Started changes for advanced parameters in types.
[Details]
This commit is the first of a few that add the feature which allows
users to include parameters for advanced types in the `-t` or `--types`
command line option. This, when complete, will allow multiple instances
of an advanced type to be run with different parameters in the same
FLITSR execution.

[File's changed]
	modified:   flitsr/advanced/__init__.py
	modified:   flitsr/advanced/flitsr.py
	modified:   flitsr/args.py