Fixed the equality checks for Groups and Elements
[Details]
Changed the functionality of the equality checks and hash functions for
spectral Groups and Elements. Before Elements used the index to check
for equality and hashing, this is now changed to check for equality
based on the name. The groups equality was also changed to not use the
hash function, which could sometimes be equal when the groups are not.

[Note]
Also fixed up using slots for the variables in Groups and Elements.

[File's changed]
	modified:   spectrum.py