Fixed Union advanced type parameter handling.
[Details]
The previous implementation for advanced type parameter handling did
not consider using python typing constructs such as Union, Optional and
Generic. These would have thus caused errors if any of the advanced
types used them in a certain way. To fix this, functionality was added
to properly deal with these types in a natural way. For instance, if an
advanced type has a parameter with type Optional[int], FLITSR will
consider this a type conversion to int.

[Note]
Also added minor updates to the documentation on how to define methods
for custom type conversions.

[File's changed]
	modified:   docs/source/adv_type_ext.rst
	modified:   flitsr/args.py