Working argument reading for advanced params.
[Details]
The reading and type conversions for parameters for advanced types
using the -t/--types argument for FLITSR is now working. The FLITSR
argument handling will now accept parameters for the -t/--types argument
in the following format: "<adv type name>(<parameter>=<value>,...)".

[Note]
Although the argument input part is finished, the use of these arguments
in FLITSR still needs to be implemented. Thus advanced parameters are
still not fully functional in FLITSR yet.

[File's changed]
	modified:   args.py