Fixed critical multiprocessing bug.
[Bugs]
Fixed a critical bug in the use of the multiprocessing module within the
run_all script. The bug was caused by the fact that when using a pool
of worker processes, the processes are reused between tasks. This was
previously not an issue, as all necessary objects were re-initialized by
FLITSR on each task execution. However a recent update to the Args
object meant that it fully relies on its Singleton design pattern, and
only initializes an Args object once in every process. This caused
the reused processes to reuse the arguments of previous runs.

The fix for this was trivial; the maximum tasks per child was set to 1,
so that processes are not reused between tasks. This may be slightly
more inefficient, as new processes need to be spawned for every FLITSR
task, but it ensures full encapsulation for each FLITSR instance. An
alternate solution would be to ensure that FLITSR re-initializes every
thing it uses on each call to the main method. This would require
re-initialization of the Args singleton on-demand.

[File's changed]
	modified:   flitsr/run_all.py