Fixed the Spectrum reset method for buckets.
[Bugs]
Fixed a bug in the Spectrum reset method when resetting buckets with a
name that has another bucket name as a prefix. In some of these cases,
the method would erroneously reset the bucket whose name is a prefix.
The bug is due to a string being an instance of an Iterable in Python,
and therefore not distinguishable from a collection type. When failing
to match the prefix bucket name to the one given, the method would check
if the given name was an Iterable, and if so, whether the prefix name is
"in" this Iterable. Since both conditions are True for the prefix bucket
name, that bucket would be reset. This was fixed by checking only for
the List type and not Iterable, as the type hint of the function
suggests.

[File's changed]
	modified:   flitsr/spectrum.py