Updated Tie documentation & fixed locations bug.
[Details]
Updated the documentation for the Tie class to reflect the addition of
active fault locations. Renamed the `fault_groups` method to
`active_fault_locations` to be more clear. The `fault_groups` name is
kept but marked as deprecated.

[Note]
The `Deprecated` package was added to the list of required packages,
which is used for the deprecation of the `fault_groups` method.

[Bugs]
- Fixed a bug in the Tie class caused by not dealing with active and
  non-active fault locations. An active fault location is a location of
  a fault that is seen for the first time in this particular tie, and
  non-active for a fault that is seen before in a previous tie.
  Non-active fault locations aren't counted for wasted effort, but add
  nothing to the number of faults found, whereas active fault locations
  do. This was fixed by counting the active locations in addition to the
  total fault locations, and storing only the active ones in the
  dictionaries used in the Tie class.

[File's changed]
	modified:   flitsr/precision_recall.py
	modified:   flitsr/tie.py
	modified:   pyproject.toml