Updated the Top1 calculations.
[Details]
Updated the Top1 calculations according to the updated theoretical
definitions. You can now specify any Top-N calculations, where N is the
cut-off threshold. There are two Top-N calculations, one-Top-N and
all-Top-N. The first returns 1.0 if at least one fault is seen in the
top N elements, or the fraction of faults seen otherwise, whereas
all-Top-N returns the total number of faults seen in the top N elements.
Both methods now use the same calculation as for precision and recall to
determine the number of faults before the cut-off N.

[Note]
This commit changes the API for FLITSR when using the Top-N calculations
and is thus a breaking change.

[File's changed]
	modified:   flitsr/args.py
	modified:   flitsr/main.py
	modified:   flitsr/precision_recall.py
	modified:   flitsr/top.py