Added summing to the merge script.
[Details]
Added a new feature to the merge script which adds a CLI option to
specify which metrics should only be summed and not averaged which
computing the aggregate. This is useful for the Top-N calculations,
which are usually reported as a sum of the bugs found in each version.

[Bugs]
Also fixed a bug in the merge script which was causing the script to
silently stop reading in results files which have empty blocks when
getting to the empty block. This is now changed to report an error
detailing the number of empty blocks found.

[File's changed]
	modified:   flitsr/merge.py