Added FLITSR* cut-off option.
[Details]
Added an option for the FLITSR*/Multi algorithm to include a cut-off
point for the number of bases to calculate before stopping. The default
for this option is no cut-off, so that the algorithm computes all
subsequent bases as usual. Note that this option is different to the
--cutoff-strategy option in that it reduces the computations of bases,
whereas the other only reduces the number of bases considered for output
and calculations. This new option is thus useful to save computational
time for cases where computing all bases would be impractical.

[File's changed]
	modified:   flitsr/advanced/flitsr.py